+++ test/runtests.jl
@@ -1,4 +1,7 @@
 using Test
 using PlutoDevMacros
+push!(LOAD_PATH, normpath(@__DIR__, "./TestPackage"))
+using TestPackage
+pop!(LOAD_PATH)
 
-include("fromparent.jl")
+include("frompackage.jl")
